Split PrecompiledScriptPluginIntegrationTest class
It was starting to get in the way of parallel test execution
